chore: rename dockerArchiver to projectArchiver
