+++ lib/goals/build.js
@@ -21,14 +21,14 @@
 // This file is responsible for creation of the archive of the application code
 // creating the BuildConfig, Imagestream and then pushing that all up
 
-const dockerArchiver = require('../docker-archiver');
+const projectArchiver = require('../project-archiver');
 const buildConfigurator = require('../build-config');
 const imageStreamConfigurator = require('../image-stream');
 const binaryBuild = require('../binary-build');
 
 module.exports = async function build (config) {
   // arvhice the application source
-  await dockerArchiver.archiveAndTar(config);
+  await projectArchiver.archiveAndTar(config);
 
   // create or update the build config
   await buildConfigurator.createOrUpdateBuildConfig(config);
@@ -37,5 +37,5 @@ module.exports = async function build (config) {
   await imageStreamConfigurator.createOrUpdateImageStream(config);
 
   // push to the cluster
-  return binaryBuild(config, `${config.projectLocation}/${dockerArchiver.DEFAULT_BUILD_LOCATION}/archive.tar`);
+  return binaryBuild(config, `${config.projectLocation}/${projectArchiver.DEFAULT_BUILD_LOCATION}/archive.tar`);
 };
