+++ lib/project-archiver.js
